# app/temp_files/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coffee point",
    "description": "This backend application is built on FastAPI and implements the full logic of cafe management.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/check-lists/end/": {
      "get": {
        "tags": [
          "check_lists"
        ],
        "summary": "Get End Check List",
        "operationId": "get_end_check_list_check_lists_end__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/items/": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Read Items",
        "operationId": "read_items_items__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ItemOut"
                  },
                  "type": "array",
                  "title": "Response Read Items Items  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "items"
        ],
        "summary": "Create Item",
        "operationId": "create_item_items__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/items/{item_id}/": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Read Item",
        "operationId": "read_item_items__item_id___get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "items"
        ],
        "summary": "Update Item",
        "operationId": "update_item_items__item_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/recipe_items/": {
      "get": {
        "tags": [
          "recipe_items"
        ],
        "summary": "Get Recipe Items",
        "operationId": "get_recipe_items_recipe_items__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecipeItemOut"
                  },
                  "type": "array",
                  "title": "Response Get Recipe Items Recipe Items  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "recipe_items"
        ],
        "summary": "Create Recipe Item",
        "operationId": "create_recipe_item_recipe_items__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/recipe_items/{recipe_item_id}/": {
      "get": {
        "tags": [
          "recipe_items"
        ],
        "summary": "Get Recipe Item",
        "operationId": "get_recipe_item_recipe_items__recipe_item_id___get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recipe_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recipe Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "recipe_items"
        ],
        "summary": "Update Recipe Item",
        "operationId": "update_recipe_item_recipe_items__recipe_item_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recipe_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recipe Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shifts/": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Get Shifts",
        "operationId": "get_shifts_shifts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftOut"
                  },
                  "type": "array",
                  "title": "Response Get Shifts Shifts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/shifts/active-shifts/": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Get Active Shifts",
        "operationId": "get_active_shifts_shifts_active_shifts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftOut"
                  },
                  "type": "array",
                  "title": "Response Get Active Shifts Shifts Active Shifts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/shifts/shift-start-update/{shift_id}/": {
      "put": {
        "tags": [
          "shifts"
        ],
        "summary": "Update Shift Start",
        "operationId": "update_shift_start_shifts_shift_start_update__shift_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/store_items/": {
      "get": {
        "tags": [
          "store_items"
        ],
        "summary": "Get Store Items",
        "operationId": "get_store_items_store_items__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StoreItemOut"
                  },
                  "type": "array",
                  "title": "Response Get Store Items Store Items  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "store_items"
        ],
        "summary": "Create Store Item",
        "operationId": "create_store_item_store_items__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/store_items/{store_item_id}/": {
      "get": {
        "tags": [
          "store_items"
        ],
        "summary": "Get Store Item",
        "operationId": "get_store_item_store_items__store_item_id___get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Store Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "store_items"
        ],
        "summary": "Update Store Item",
        "operationId": "update_store_item_store_items__store_item_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Store Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckListOut": {
        "properties": {
          "check_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Check List"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "check_list",
          "id"
        ],
        "title": "CheckListOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "measurement": {
            "type": "string",
            "title": "Measurement"
          }
        },
        "type": "object",
        "required": [
          "name",
          "measurement"
        ],
        "title": "ItemCreate"
      },
      "ItemOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "measurement": {
            "type": "string",
            "title": "Measurement"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "measurement",
          "id",
          "active"
        ],
        "title": "ItemOut"
      },
      "ItemUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "measurement": {
            "type": "string",
            "title": "Measurement"
          }
        },
        "type": "object",
        "required": [
          "name",
          "measurement"
        ],
        "title": "ItemUpdate"
      },
...
      "RecipeItemCreate": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "item_id",
          "amount"
        ],
        "title": "RecipeItemCreate"
      },
      "RecipeItemOut": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "item_id",
          "amount",
          "id"
        ],
        "title": "RecipeItemOut"
      },
      "RecipeItemUpdate": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "item_id",
          "amount"
        ],
        "title": "RecipeItemUpdate"
      },
...
      "ShiftOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "start_time": {
            "title": "Start Time",
            "type": [
              "string",
              "null"
            ]
          },
          "end_time": {
            "title": "End Time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "active"
        ],
        "title": "ShiftOut"
      },
      "StoreItemCreate": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "supply_id": {
            "type": "string",
            "format": "uuid",
            "title": "Supply Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "price_per_item": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Price Per Item"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "supply_id",
          "amount",
          "price_per_item"
        ],
        "title": "StoreItemCreate"
      },
      "StoreItemOut": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "supply_id": {
            "type": "string",
            "format": "uuid",
            "title": "Supply Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "price_per_item": {
            "type": "string",
            "title": "Price Per Item"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "supply_id",
          "amount",
          "price_per_item",
          "id"
        ],
        "title": "StoreItemOut"
      },
      "StoreItemUpdate": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "supply_id": {
            "type": "string",
            "format": "uuid",
            "title": "Supply Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "price_per_item": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Price Per Item"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "supply_id",
          "amount",
          "price_per_item"
        ],
        "title": "StoreItemUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}